// naomi-seeds-backend/src/modules/orders/orders.enum.ts
export enum OrderStatus {
    PENDING = "PENDING",
    APPROVED = "APPROVED",
    NOT_APPLICABLE = "NOT_APPLICABLE",
    SHIPPED = "SHIPPED",
    OUT_FOR_DELIVERY = "OUT_FOR_DELIVERY",
    REJECT = "REJECT",
  }


  export enum TypeOfSale {
    COMPANY_SALE = 'COMPANY_SALE',
    IPT = 'IPT',
    SALES_RETURN = 'SALES_RETURN',
  }